Remove implicit promotion compile warning

+++ driver/ntv2genlock2.c
@@ -439,7 +439,7 @@ static bool configure_genlock2(struct ntv2_genlock2 *ntv2_gen, struct ntv2_genlo
 		if (check)
 		{
 			uint8_t readBytes[256];
-            uint32_t i;
+            uint16_t i;
 			if (spi_genlock2_read(ntv2_gen, gdat->addr, readBytes, gdat->size))
 			{
 				for (i = 0; i < gdat->size; i++)
